Fixed notifications.  Fixed On/Off buttons. About to break app trying to fix threading issues.

+++ Shoutbreak/src/com/shoutbreak/service/CellDensity.java
@@ -4,7 +4,7 @@ public class CellDensity {
 
 	public int cellX = -1;
 	public int cellY = -1;
-	public float density = -1;
+	public double density = -1;
 	public boolean isSet = false;
 	public String lastUpdated;
